% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_bilders.R
\name{create_reports}
\alias{create_reports}
\title{Create the samples infos to create the Report}
\usage{
create_reports(file_info, file_genos)
}
\arguments{
\item{file_info}{Get patients' file info form web_app generated through Prepare_data_Mac_v2.py}

\item{file_genos}{Get patients' genotype info file form web_app generated through Prepare_data_Mac_v2.py}
}
\description{
Create the samples infos to create the Report
}
