% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_drug_infos.R
\name{extract_gene_guide}
\alias{extract_gene_guide}
\title{Extract gene specific drug guide}
\usage{
extract_gene_guide(
  feature,
  drug_guide,
  sample_genotype,
  fvar = c("Gene", "Variant"),
  pg = c("Genotype", "Phenotype")
)
}
\arguments{
\item{fvar}{feature Variable name to select column accordingly}

\item{pg}{Phenotype or Genotype}
}
\description{
Extract gene specific drug guide
}
