% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidelines.R
\docType{data}
\name{fcinetica}
\alias{fcinetica}
\title{Data from FarmacoCinetica}
\format{
A data frame
\describe{
  \item{Drug}{Farmaco}
  \item{Gene}{Gene interessato}
  \item{Phenotype}{Fenotipo del gene}
  \item{Category}{Categoria dell'indicazione}
  \item{Indicazione_Clinica}{Indicazioni clinica testuale}
}
}
\usage{
fcinetica
}
\description{
Guidelines for farmacinetica
}
\keyword{datasets}
