% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidelines.R
\docType{data}
\name{fdinamica}
\alias{fdinamica}
\title{Data from FarmacoDinamica}
\format{
A data frame
\describe{
  \item{Drug}{Farmaco}
  \item{Gene}{Gene interessato}
  \item{Variant}{Variante interessata}
  \item{Genotype}{Genotype del paziente}
  \item{Category}{Categoria dell'indicazione}
  \item{Indicazione_Clinica}{Indicazioni clinica testuale}
}
}
\usage{
fdinamica
}
\description{
Guidelines for farmacinetica
}
\keyword{datasets}
