% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instructions.R
\name{get_patient_clinical_guidelines}
\alias{get_patient_clinical_guidelines}
\title{Read and format sample info and sample genotypes}
\usage{
get_patient_clinical_guidelines(patient, sample_info_table)
}
\description{
Read and format sample info and sample genotypes
}
