% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_info_table.R
\name{read_sample_genos}
\alias{read_sample_genos}
\title{Read in sample genotypes}
\usage{
read_sample_genos(file)
}
\arguments{
\item{file}{file to read in}
}
\description{
Read in sample genotypes
}
