% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_info_table.R
\name{generate_sample_info_table}
\alias{generate_sample_info_table}
\title{Read and format sample info and sample genotypes}
\usage{
generate_sample_info_table(sample_info_file, sample_genotypes_file)
}
\description{
Read and format sample info and sample genotypes
}
