% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instructions.R
\name{get_patient_clinical_guidelines}
\alias{get_patient_clinical_guidelines}
\title{Read and format sample info and sample genotypes}
\usage{
get_patient_clinical_guidelines(
  patient,
  sample_info_table,
  remove_last = FALSE
)
}
\arguments{
\item{patient}{patient id}

\item{sample_info_table}{table obtained merging sample_info and sample_geno}

\item{remove_last}{internal use only: remove the last known drug}
}
\description{
Read and format sample info and sample genotypes
}
